renames with kebab case
