+++ src/app/page.tsx
@@ -1,11 +1,11 @@
-import Header from '@/components/Header'
-import Hero from '@/components/Hero'
-import Skills from '@/components/Skills'
-import ProjectCarousel from '@/components/ProjectsCarousel'
-import GitHubContributions from '@/components/GitHubContributions'
-import Experience from '@/components/Experience'
-import Activities from '@/components/Activities'
-import Footer from '@/components/Footer'
+import Header from '@/components/header'
+import Hero from '@/components/hero'
+import Skills from '@/components/skills'
+import ProjectCarousel from '@/components/project-carousel'
+import GitHubContributions from '@/components/github-contributions'
+import Experience from '@/components/experience'
+import Activities from '@/components/activities'
+import Footer from '@/components/footer'
 // import ChatBot from '@/components/ChatBot'
 
 export default function Home() {
