+++ src/components/project-carousel.tsx
